+++ cmd/frontend/graphqlbackend/user.go
@@ -305,7 +305,7 @@ func viewerCanChangeUsername(ctx context.Context, userID int32) bool {
 	if err := backend.CheckSiteAdminOrSameUser(ctx, userID); err != nil {
 		return false
 	}
-	if conf.Get().Critical.AuthEnableUsernameChanges {
+	if conf.Get().AuthEnableUsernameChanges {
 		return true
 	}
 	// 🚨 SECURITY: Only site admins are allowed to change a user's username when auth.enableUsernameChanges == false.