+++ cmd/frontend/internal/app/ui/handlers.go
@@ -98,10 +98,10 @@ func repoShortName(name api.RepoName) string {
 func newCommon(w http.ResponseWriter, r *http.Request, title string, serveError func(w http.ResponseWriter, r *http.Request, err error, statusCode int)) (*Common, error) {
 	common := &Common{
 		Injected: InjectedHTML{
-			HeadTop:    template.HTML(conf.Get().Critical.HtmlHeadTop),
-			HeadBottom: template.HTML(conf.Get().Critical.HtmlHeadBottom),
-			BodyTop:    template.HTML(conf.Get().Critical.HtmlBodyTop),
-			BodyBottom: template.HTML(conf.Get().Critical.HtmlBodyBottom),
+			HeadTop:    template.HTML(conf.Get().HtmlHeadTop),
+			HeadBottom: template.HTML(conf.Get().HtmlHeadBottom),
+			BodyTop:    template.HTML(conf.Get().HtmlBodyTop),
+			BodyBottom: template.HTML(conf.Get().HtmlBodyBottom),
 		},
 		Context:  jscontext.NewJSContextFromRequest(r),
 		AssetURL: assetsutil.URL("").String(),