+++ cmd/server/shared/globals.go
@@ -13,7 +13,6 @@ var SrcProfServices = []map[string]string{
 	{"Name": "frontend", "Host": "127.0.0.1:6063"},
 	{"Name": "gitserver", "Host": "127.0.0.1:6068"},
 	{"Name": "searcher", "Host": "127.0.0.1:6069"},
-	{"Name": "management-console", "Host": "127.0.0.1:6075"},
 	{"Name": "symbols", "Host": "127.0.0.1:6071"},
 	{"Name": "repo-updater", "Host": "127.0.0.1:6074"},
 	{"Name": "query-runner", "Host": "127.0.0.1:6067"},