+++ enterprise/cmd/frontend/auth/httpheader/config.go
@@ -9,7 +9,7 @@ import (
 // site config; if there is more than 1, it returns multiple == true (which the caller should handle
 // by returning an error and refusing to proceed with auth).
 func getProviderConfig() (pc *schema.HTTPHeaderAuthProvider, multiple bool) {
-	for _, p := range conf.Get().Critical.AuthProviders {
+	for _, p := range conf.Get().AuthProviders {
 		if p.HttpHeader != nil {
 			if pc != nil {
 				return pc, true // multiple http-header auth providers
@@ -26,13 +26,13 @@ func init() {
 
 func validateConfig(c conf.Unified) (problems conf.Problems) {
 	var httpHeaderAuthProviders int
-	for _, p := range c.Critical.AuthProviders {
+	for _, p := range c.AuthProviders {
 		if p.HttpHeader != nil {
 			httpHeaderAuthProviders++
 		}
 	}
 	if httpHeaderAuthProviders >= 2 {
-		problems = append(problems, conf.NewCriticalProblem(`at most 1 http-header auth provider may be used`))
+		problems = append(problems, conf.NewSiteProblem(`at most 1 http-header auth provider may be used`))
 	}
 	return problems
 }