+++ enterprise/cmd/frontend/internal/registry/http_api.go
@@ -75,7 +75,7 @@ func toRegistryAPIExtension(ctx context.Context, v *dbExtension) (*registry.Exte
 		return nil, err
 	}
 
-	baseURL := strings.TrimSuffix(conf.Get().Critical.ExternalURL, "/")
+	baseURL := strings.TrimSuffix(conf.Get().ExternalURL, "/")
 	return &registry.Extension{
 		UUID:        v.UUID,
 		ExtensionID: v.NonCanonicalExtensionID,