+++ web/src/site-admin/SiteAdminOverviewPage.tsx
@@ -14,7 +14,6 @@ import { queryGraphQL } from '../backend/graphql'
 import { Collapsible } from '../components/Collapsible'
 import { PageTitle } from '../components/PageTitle'
 import { eventLogger } from '../tracking/eventLogger'
-import { SiteAdminManagementConsolePassword } from './SiteAdminManagementConsolePassword'
 import { UsageChart } from './SiteAdminUsageStatisticsPage'
 import { ErrorAlert } from '../components/alerts'
 
@@ -108,9 +107,6 @@ export class SiteAdminOverviewPage extends React.Component<Props, State> {
         return (
             <div className="site-admin-overview-page py-3">
                 <PageTitle title="Overview - Admin" />
-                <div className="mb-3">
-                    <SiteAdminManagementConsolePassword />
-                </div>
                 {this.props.overviewComponents.length > 0 && (
                     <div className="mb-4">
                         {this.props.overviewComponents.map((C, i) => (