+++ web/src/site-admin/backend.tsx
@@ -316,12 +316,10 @@ type SettingsSubject = Pick<GQL.SettingsSubject, 'settingsURL' | '__typename'> &
 }
 
 /**
- * All configuration and settings in one place. Excludes critical site config for now, because this
- * must be fetched from management console
+ * All configuration and settings in one place.
  */
 export interface AllConfig {
     site: GQL.ISiteConfiguration
-    critical: GQL.ICriticalConfiguration
     externalServices: Partial<Record<GQL.ExternalServiceKind, ExternalServiceConfig>>
     settings: {
         subjects: SettingsSubject[]
@@ -339,10 +337,6 @@ export function fetchAllConfigAndSettings(): Observable<AllConfig> {
                 site {
                     id
                     configuration {
-                        id
-                        effectiveContents
-                    }
-                    criticalConfiguration {
                         id
                         effectiveContents
                     }
@@ -417,10 +411,6 @@ export function fetchAllConfigAndSettings(): Observable<AllConfig> {
                     data.site.configuration &&
                     data.site.configuration.effectiveContents &&
                     parseJSONC(data.site.configuration.effectiveContents),
-                critical:
-                    data.site &&
-                    data.site.criticalConfiguration &&
-                    parseJSONC(data.site.criticalConfiguration.effectiveContents),
                 externalServices,
                 settings: {
                     subjects: settingsSubjects,